Rename files from 'duplicate' to 'issue'
