+++ src/api/types.ts
@@ -1,5 +1,5 @@
-import { Issue } from '../duplicate-results/types';
-import { IssueSortOption, IssueStatusFilter } from '../duplicate-search/types';
+import { Issue } from '../issue-search-results/types';
+import { IssueSortOption, IssueStatusFilter } from '../issue-search/types';
 import { LearnMoreLink, TaskDetails } from '../static-data/reporting-config/types';
 
 /**
