+++ src/issue-search-results/sub-components/use-show-banner.ts
@@ -3,13 +3,13 @@ import { useAppSelector } from '../../app/hooks';
 import {
 	selectDuplicateRequestsWereMade,
 	selectDuplicateResultsRequestStatus,
-} from '../duplicate-results-slice';
+} from '../issue-search-results-slice';
 
 export function useShowBanner() {
 	const resultsRequestStatus = useAppSelector( selectDuplicateResultsRequestStatus );
 	const requestsWereMade = useAppSelector( selectDuplicateRequestsWereMade );
 
-	// We want to handle the case when we come back to the duplicate searching page.
+	// We want to handle the case when we come back to the issue searching page.
 	// Since this is local state, we want to tie it back to the redux state.
 	const [ showBanner, setShowBanner ] = useState( requestsWereMade );
